v. 0.5.1 Исправлен отображение фото и описания для SE
