+++ Course2FinalTask/Course2FinalTask/Controller/FeedCell/FeedCollectionViewCell.xib
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="15705" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_0" orientation="portrait" appearance="light"/>
+    <device id="retina5_5" orientation="portrait" appearance="light"/>
     <dependencies>
         <deployment identifier="iOS"/>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="15706"/>
@@ -10,7 +10,7 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="FeedViewController" customModule="Course2FinalTask" customModuleProvider="target"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" reuseIdentifier="Cell" id="gTV-IL-0wX" customClass="DetailCollectionViewCell" customModule="Course2FinalTask" customModuleProvider="target">
+        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" reuseIdentifier="Cell" id="gTV-IL-0wX" customClass="FeedCollectionViewCell" customModule="Course2FinalTask" customModuleProvider="target">
             <rect key="frame" x="0.0" y="0.0" width="320" height="470"/>
             <autoresizingMask key="autoresizingMask"/>
             <view key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center">
@@ -22,6 +22,9 @@
                     </imageView>
                     <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="SfU-eq-VZD">
                         <rect key="frame" x="0.0" y="51" width="320" height="320"/>
+                        <constraints>
+                            <constraint firstAttribute="width" secondItem="SfU-eq-VZD" secondAttribute="height" multiplier="1:1" id="PsS-Yf-lCw"/>
+                        </constraints>
                     </imageView>
                     <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="zF1-Ni-sRJ">
                         <rect key="frame" x="261" y="372" width="44" height="44"/>
@@ -37,13 +40,13 @@
                         <rect key="frame" x="58" y="8" width="247" height="35"/>
                         <subviews>
                             <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="userName" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="gmS-t9-oZc">
-                                <rect key="frame" x="0.0" y="0.0" width="247" height="17.5"/>
+                                <rect key="frame" x="0.0" y="0.0" width="247" height="17.666666666666668"/>
                                 <fontDescription key="fontDescription" type="system" pointSize="14"/>
                                 <nil key="textColor"/>
                                 <nil key="highlightedColor"/>
                             </label>
                             <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="data" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="pQF-jY-tOg">
-                                <rect key="frame" x="0.0" y="17.5" width="247" height="17.5"/>
+                                <rect key="frame" x="0.0" y="17.666666666666671" width="247" height="17.333333333333329"/>
                                 <fontDescription key="fontDescription" type="system" pointSize="14"/>
                                 <nil key="textColor"/>
                                 <nil key="highlightedColor"/>
@@ -55,7 +58,7 @@
                         </constraints>
                     </stackView>
                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Likes:" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="5" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="d3N-yl-MXw">
-                        <rect key="frame" x="15" y="379" width="37.5" height="24"/>
+                        <rect key="frame" x="15.000000000000004" y="379" width="37.333333333333343" height="17"/>
                         <fontDescription key="fontDescription" type="system" pointSize="14"/>
                         <nil key="textColor"/>
                         <nil key="highlightedColor"/>
@@ -70,7 +73,7 @@
             </view>
             <constraints>
                 <constraint firstItem="d3N-yl-MXw" firstAttribute="leading" secondItem="ZTg-uK-7eu" secondAttribute="leading" constant="15" id="7nx-d9-EYr"/>
-                <constraint firstItem="ZTg-uK-7eu" firstAttribute="bottom" secondItem="zAj-tk-cU6" secondAttribute="bottom" constant="29" id="94p-yN-fMH"/>
+                <constraint firstItem="ZTg-uK-7eu" firstAttribute="bottom" relation="lessThanOrEqual" secondItem="zAj-tk-cU6" secondAttribute="bottom" constant="29" id="94p-yN-fMH"/>
                 <constraint firstItem="SfU-eq-VZD" firstAttribute="top" secondItem="mxu-1l-XIR" secondAttribute="bottom" constant="8" id="Fg9-VG-asa"/>
                 <constraint firstItem="GID-fZ-GW7" firstAttribute="leading" secondItem="mxu-1l-XIR" secondAttribute="trailing" constant="8" id="HfT-Gv-O8V"/>
                 <constraint firstItem="ZTg-uK-7eu" firstAttribute="trailing" secondItem="zAj-tk-cU6" secondAttribute="trailing" constant="15" id="K09-he-QRv"/>
@@ -78,7 +81,6 @@
                 <constraint firstItem="mxu-1l-XIR" firstAttribute="top" secondItem="ZTg-uK-7eu" secondAttribute="top" constant="8" id="KqY-Y4-lAf"/>
                 <constraint firstAttribute="trailing" secondItem="SfU-eq-VZD" secondAttribute="trailing" id="LIB-du-ga0"/>
                 <constraint firstItem="GID-fZ-GW7" firstAttribute="top" secondItem="gTV-IL-0wX" secondAttribute="top" constant="8" id="MWe-d0-8F9"/>
-                <constraint firstItem="zAj-tk-cU6" firstAttribute="top" secondItem="d3N-yl-MXw" secondAttribute="bottom" constant="21" id="Mu5-gQ-R9w"/>
                 <constraint firstItem="d3N-yl-MXw" firstAttribute="top" secondItem="SfU-eq-VZD" secondAttribute="bottom" constant="8" id="NdT-I5-vuk"/>
                 <constraint firstItem="SfU-eq-VZD" firstAttribute="leading" secondItem="gTV-IL-0wX" secondAttribute="leading" id="Phh-hR-EiA"/>
                 <constraint firstItem="mxu-1l-XIR" firstAttribute="leading" secondItem="ZTg-uK-7eu" secondAttribute="leading" constant="15" id="Rbl-Lm-K7d"/>
